# docs/swagger.yaml
definitions:
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  models.Role:
    enum:
    - admin
    - user
    - lecturer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
    - RoleLecturer
...
  users.UpdateUserRequest:
    properties:
      id:
        type: string
      user_fullname:
        type: string
      user_role:
        $ref: '#/definitions/models.Role'
      year:
        type: integer
    required:
    - id
    type: object
info:
  contact: {}
paths:
...
  /user/update:
    post:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: UpdateUserRequest
        in: body
        name: UpdateUserRequest
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update user
      tags:
      - User
swagger: "2.0"